+++ apps/Cargo.toml
@@ -53,7 +53,6 @@ config = "0.11.0"
 ed25519-dalek = {version = "1.0.1", default-features = false, features = ["rand", "u64_backend", "serde"]}
 eyre = "0.6.5"
 futures = "0.3"
-hex = "0.4.2"
 libp2p = "0.38.0"
 prost = "0.7.0"
 prost-types = "0.7.0"
@@ -88,3 +87,6 @@ anoma_shared = {path = "../shared", features = ["testing", "wasm-runtime"]}
 cargo-watch = "7.5.0"
 proptest = "1.0.0"
 tempfile = "3.2.0"
+
+[build-dependencies]
+tonic-build = "0.4.0"
