+++ apps/src/bin/anoma-client/cli.rs
@@ -2,9 +2,9 @@ use std::fs::File;
 use std::io::Write;
 
 use anoma::client::tx;
+use anoma::proto::services::rpc_service_client::RpcServiceClient;
+use anoma::proto::{services, RpcMessage};
 use anoma::{cli, wallet};
-use anoma_shared::proto::services::rpc_service_client::RpcServiceClient;
-use anoma_shared::proto::{self, services, RpcMessage};
 use anoma_shared::types::intent::Intent;
 use anoma_shared::types::key::ed25519::Signed;
 use anoma_shared::types::{token, Address};
@@ -97,7 +97,7 @@ pub async fn main() -> Result<()> {
 async fn gossip_intent(node_addr: String, data_path: String, topic: String) {
     let mut client = RpcServiceClient::connect(node_addr).await.unwrap();
     let data = std::fs::read(data_path).expect("data file IO error");
-    let intent = proto::Intent::new(data);
+    let intent = anoma_shared::proto::Intent::new(data);
     let message: services::RpcMessage =
         RpcMessage::new_intent(intent, topic).into();
     let response = client
