+++ apps/src/lib/mod.rs
@@ -7,6 +7,7 @@ pub mod genesis;
 pub mod gossiper;
 pub mod logging;
 pub mod node;
+pub mod proto;
 pub mod types;
 pub mod wallet;
