+++ apps/src/lib/node/gossip/mod.rs
@@ -3,12 +3,12 @@ mod network_behaviour;
 mod p2p;
 mod rpc;
 
-use anoma_shared::proto::services::{rpc_message, RpcResponse};
 use thiserror::Error;
 use tokio::sync::{mpsc, oneshot};
 
 use self::p2p::P2P;
 use crate::config::IntentGossiper;
+use crate::proto::services::{rpc_message, RpcResponse};
 use crate::types::MatchmakerMessage;
 
 #[derive(Error, Debug)]
