+++ apps/src/lib/node/gossip/rpc.rs
@@ -1,14 +1,15 @@
 use std::net::SocketAddr;
 use std::thread;
 
-use anoma_shared::proto::services::rpc_service_server::{
-    RpcService, RpcServiceServer,
-};
-use anoma_shared::proto::services::{rpc_message, RpcMessage, RpcResponse};
 use tokio::sync::mpsc::{self, Sender};
 use tokio::sync::oneshot;
 use tonic::transport::Server;
 use tonic::{Request as TonicRequest, Response as TonicResponse, Status};
+
+use crate::proto::services::rpc_service_server::{
+    RpcService, RpcServiceServer,
+};
+use crate::proto::services::{rpc_message, RpcMessage, RpcResponse};
 
 #[derive(Debug)]
 struct Rpc {
