+++ apps/src/lib/proto/generated.rs
@@ -1,0 +1,1 @@
+pub mod services;
