+++ apps/src/lib/proto/mod.rs
@@ -1,0 +1,5 @@
+mod generated;
+mod types;
+
+pub use generated::services;
+pub use types::{IntentMessage, RpcMessage, SubscribeTopicMessage};
