+++ proto/services.proto
@@ -8,7 +8,6 @@ service RPCService {
   rpc SendMessage(RpcMessage) returns (RpcResponse);
 }
 
-
 message IntentMessage{
   types.Intent intent = 1;
   string topic = 2;
@@ -23,7 +22,6 @@ message RpcMessage {
     IntentMessage intent = 1;
     SubscribeTopicMessage topic = 2;
     types.Dkg dkg = 3;
-
   }
 }
