+++ shared/src/proto/generated.rs
@@ -1,2 +1,1 @@
-pub mod services;
 pub mod types;
