+++ shared/src/proto/types.rs
@@ -6,7 +6,7 @@ use prost::Message;
 use prost_types::Timestamp;
 use thiserror::Error;
 
-use super::generated::{services, types};
+use super::generated::types;
 
 #[derive(Error, Debug)]
 pub enum Error {
@@ -44,7 +44,7 @@ impl TryFrom<&[u8]> for Tx {
         };
         Ok(Tx {
             code: tx.code,
-            data: tx.data,
+            data: tx.data.map(|tx_data| tx_data.data),
             timestamp,
         })
     }
@@ -54,7 +54,7 @@ impl From<Tx> for types::Tx {
     fn from(tx: Tx) -> Self {
         types::Tx {
             code: tx.code.clone(),
-            data: tx.data.clone(),
+            data: tx.data.map(|data| types::TxData { data }),
             timestamp: Some(tx.timestamp),
         }
     }
@@ -174,105 +174,6 @@ impl DkgGossipMessage {
     }
 }
 
-pub enum RpcMessage {
-    IntentMessage(IntentMessage),
-    SubscribeTopicMessage(SubscribeTopicMessage),
-    Dkg(Dkg),
-}
-
-impl From<RpcMessage> for services::RpcMessage {
-    fn from(message: RpcMessage) -> Self {
-        let message = match message {
-            RpcMessage::IntentMessage(m) => {
-                services::rpc_message::Message::Intent(m.into())
-            }
-            RpcMessage::SubscribeTopicMessage(m) => {
-                services::rpc_message::Message::Topic(m.into())
-            }
-            RpcMessage::Dkg(d) => services::rpc_message::Message::Dkg(d.into()),
-        };
-        services::RpcMessage {
-            message: Some(message),
-        }
-    }
-}
-
-impl RpcMessage {
-    pub fn new_intent(intent: Intent, topic: String) -> Self {
-        RpcMessage::IntentMessage(IntentMessage::new(intent, topic))
-    }
-
-    pub fn new_topic(topic: String) -> Self {
-        RpcMessage::SubscribeTopicMessage(SubscribeTopicMessage::new(topic))
-    }
-
-    pub fn new_dkg(dkg: Dkg) -> Self {
-        RpcMessage::Dkg(dkg)
-    }
-}
-
-#[derive(Debug, PartialEq)]
-pub struct IntentMessage {
-    pub intent: Intent,
-    pub topic: String,
-}
-
-impl TryFrom<services::IntentMessage> for IntentMessage {
-    type Error = Error;
-
-    fn try_from(message: services::IntentMessage) -> Result<Self> {
-        match message.intent {
-            Some(intent) => Ok(IntentMessage {
-                intent: intent.try_into()?,
-                topic: message.topic,
-            }),
-            None => Err(Error::NoIntentError),
-        }
-    }
-}
-
-impl From<IntentMessage> for services::IntentMessage {
-    fn from(message: IntentMessage) -> Self {
-        services::IntentMessage {
-            intent: Some(message.intent.into()),
-            topic: message.topic,
-        }
-    }
-}
-
-impl IntentMessage {
-    pub fn new(intent: Intent, topic: String) -> Self {
-        IntentMessage { intent, topic }
-    }
-}
-
-#[derive(Debug, PartialEq)]
-pub struct SubscribeTopicMessage {
-    pub topic: String,
-}
-
-impl From<services::SubscribeTopicMessage> for SubscribeTopicMessage {
-    fn from(message: services::SubscribeTopicMessage) -> Self {
-        SubscribeTopicMessage {
-            topic: message.topic,
-        }
-    }
-}
-
-impl From<SubscribeTopicMessage> for services::SubscribeTopicMessage {
-    fn from(message: SubscribeTopicMessage) -> Self {
-        services::SubscribeTopicMessage {
-            topic: message.topic,
-        }
-    }
-}
-
-impl SubscribeTopicMessage {
-    pub fn new(topic: String) -> Self {
-        SubscribeTopicMessage { topic }
-    }
-}
-
 #[derive(Clone, Debug, PartialEq)]
 pub struct Intent {
     pub data: Vec<u8>,
@@ -368,8 +269,8 @@ mod tests {
     #[test]
     fn test_tx() {
         let code = "wasm code".as_bytes().to_owned();
-        let data = Some("arbitrary data".as_bytes().to_owned());
-        let tx = Tx::new(code.clone(), data.clone());
+        let data = "arbitrary data".as_bytes().to_owned();
+        let tx = Tx::new(code.clone(), Some(data.clone()));
 
         let bytes = tx.to_bytes();
         let tx_from_bytes =
@@ -378,7 +279,7 @@ mod tests {
 
         let types_tx = types::Tx {
             code,
-            data,
+            data: Some(types::TxData { data }),
             timestamp: None,
         };
         let mut bytes = vec![];
@@ -411,43 +312,6 @@ mod tests {
         let message_from_bytes = DkgGossipMessage::try_from(bytes.as_ref())
             .expect("decoding failed");
         assert_eq!(message_from_bytes, message);
-    }
-
-    #[test]
-    fn test_intent_message() {
-        let data = "arbitrary data".as_bytes().to_owned();
-        let intent = Intent::new(data);
-        let topic = "arbitrary string".to_owned();
-        let intent_message = IntentMessage::new(intent.clone(), topic.clone());
-
-        let intent_rpc_message = RpcMessage::new_intent(intent, topic);
-        let services_rpc_message: services::RpcMessage =
-            intent_rpc_message.into();
-        match services_rpc_message.message {
-            Some(services::rpc_message::Message::Intent(i)) => {
-                let message_from_types =
-                    IntentMessage::try_from(i).expect("no intent");
-                assert_eq!(intent_message, message_from_types);
-            }
-            _ => panic!("no intent message"),
-        }
-    }
-
-    #[test]
-    fn test_topic_message() {
-        let topic = "arbitrary string".to_owned();
-        let topic_message = SubscribeTopicMessage::new(topic.clone());
-
-        let topic_rpc_message = RpcMessage::new_topic(topic.clone());
-        let services_rpc_message: services::RpcMessage =
-            topic_rpc_message.into();
-        match services_rpc_message.message {
-            Some(services::rpc_message::Message::Topic(t)) => {
-                let message_from_types = SubscribeTopicMessage::from(t);
-                assert_eq!(topic_message, message_from_types);
-            }
-            _ => panic!("no intent message"),
-        }
     }
 
     #[test]
